+++ src/benchmark/sssp_demo.c
@@ -140,8 +140,8 @@ int main (int argc, char **argv)
                 G, src, delta, true, msg)) ;
             LAGraph_TRY (LAGraph_Toc (&ttrial, tic, msg)) ;
 
-            printf ("sssp15:  threads: %2d trial: %2d source %9lu "
-                "time: %10.4f sec\n", nthreads, trial, src, ttrial) ;
+            printf ("sssp15:  threads: %2d trial: %2d source %g "
+                "time: %10.4f sec\n", nthreads, trial, (double) src, ttrial) ;
             total_time += ttrial ;
 
 #if LG_CHECK_RESULT
